% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.r
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{A numeric vector or integer}
}
\value{
Probability that 2 or more people in the random sample of size x have a common birthday
}
\description{
Birthday
}
\examples{
birthday(20:24)
}
