% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.r
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Contamination of Fish in the Tennessee River}
\format{
## `ddt`
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{River abbreviation}
  \item{MILE}{Miles upstream}
  \item{SPECIES}{Species name}
  \item{LENGTH}{Centimeters}
  \item{WEIGHT}{Grams}
  \item{DDT}{Ppm}
}
}
\source{
<https://github.com/MATHSTATSOU/Intro2R/tree/master/data>
}
\usage{
ddt
}
\description{
A set of data from members of the U.S. Army Corps of Engineers
}
\keyword{datasets}
