% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage data}
\format{
## `fire`
A tibble with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from fire to nearest fire station in miles}
  \item{DAMAGE}{Amount of damage in thousands of dollars}
}
}
\source{
<https://github.com/MATHSTATSOU/Intro2R/tree/master/data>
}
\usage{
fire
}
\description{
A sample of fires of large suburb inside a major city from an fire insurance
company
}
\keyword{datasets}
