% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.r
\name{myboot2}
\alias{myboot2}
\title{My Bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{An integer; number of iterations}

\item{x}{A vector; sample data to bootstrap}

\item{fun}{A string; function for desired estimated statistic}

\item{alpha}{An integer; confidence value}

\item{cx}{An integer; modifies the 'cex' param from graphical function 'par'}

\item{...}{Additional arguments to be used for the returned histogram}
}
\value{
A histogram and a list of the confidence interval, function, data, and x statistic
}
\description{
My Bootstrap
}
\examples{
myboot2(iter = 100000, x = rnorm(20, 0, 1), fun = "var", alpha = 0.01, cx = 1.2, xlab = "Variance")
}
