% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.r
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{An integer; sample size}

\item{iter}{An integer; number of samples/iterations}

\item{a}{An integer; binomial distribution parameter a}

\item{b}{An integer; binomial distribution parameter b}
}
\value{
A histogram of sample means
}
\description{
mycltu
}
\examples{
mycltu(20, 100000, a = 5, b = 15)
}
