% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.r
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimator for Normal Distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{vector of observations}

\item{mu}{vector of test values for mu (use seq())}

\item{sig}{vector of test values for sigma (use seq())}

\item{...}{graphical paramters}
}
\value{
the sample, a matrix of the indexes where the MLE was found, and the maximum likelihood
}
\description{
Maximum Likelihood Estimator for Normal Distribution
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
