% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.r
\name{myncurve}
\alias{myncurve}
\title{My N Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{An integer, mean}

\item{sigma}{An integer, sd}

\item{a}{An integer, lower tail}
}
\value{
A normal curve plot with the area of the lower tail overlayed, and the average and standard deviation as a list
}
\description{
My N Curve
}
\examples{
myncurve(0, 1, 1)
}
