% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.r
\name{mysample}
\alias{mysample}
\title{My Sample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{An integer}

\item{iter}{An integer}

\item{time}{An integer}
}
\value{
A boxplot for each iteration
}
\description{
My Sample
}
\examples{
mysample(10000)
}
