% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.r
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{An integer; number of available seats}

\item{gamma}{An integer; "pain" factor; probability of overbooking}

\item{p}{An integer; probability a passenger will show who bought a ticket}
}
\value{
2 plots and a list of values, including n for both the discrete and continuous cases, and the inputs of the function
}
\description{
ntickets
}
\examples{
ntickets(400, 0.02, 0.95)
}
